// src/api/models/linkCategory.ts
/**
 * Generated by orval v6.10.2 🍺
 * Do not edit manually.
 * API
 * API DOCS
 * OpenAPI spec version: 1.0
 */
import type { LinkCategoryLabels } from './linkCategoryLabels';

export interface LinkCategory {
  index?: number;
  enabled?: boolean;
  labels?: LinkCategoryLabels;
  links?: string[];
}